+++ projects/ngx-mat-intl-tel-input/src/lib/ngx-mat-intl-tel-input.module.ts
@@ -10,12 +10,14 @@ import { MatIconModule } from '@angular/material/icon';
 import { SearchPipe } from './search.pipe';
 import { NgxMatItlTelChipInputComponent } from './ngx-mat-itl-tel-chip-input/ngx-mat-itl-tel-chip-input.component';
 import { NgxMatItlTelInputComponent } from './ngx-mat-itl-tel-input/ngx-mat-itl-tel-input.component';
+import { NgxMatItlCountrySelectComponent } from './ngx-mat-itl-country-select/ngx-mat-itl-country-select.component';
 
 @NgModule({
   declarations: [
     NgxMatItlTelInputComponent,
     NgxMatItlTelChipInputComponent,
     SearchPipe,
+    NgxMatItlCountrySelectComponent,
   ],
   imports: [
     CommonModule,
