+++ projects/ngx-mat-intl-tel-input/src/lib/ngx-mat-itl-tel-chip-input/ngx-mat-itl-tel-chip-input.component.html
@@ -1,11 +1,10 @@
-<mat-chip-list #telList [formControl]="control">
-    <mat-chip *ngFor="let tel of control.value" [selectable]="false" [removable]="true" (removed)="onRemoveTel(tel)">
-        {{tel}}
-        <mat-icon matChipRemove>cancel</mat-icon>
-    </mat-chip>
-</mat-chip-list>
-
 <mat-form-field [appearance]="appearance">
+    <mat-chip-list #telList [formControl]="control">
+        <mat-chip *ngFor="let tel of control.value" [selectable]="false" [removable]="true" (removed)="onRemoveTel(tel)">
+            {{tel}}
+            <mat-icon matChipRemove>cancel</mat-icon>
+        </mat-chip>
+    </mat-chip-list>
     <mat-label>{{label}}</mat-label>
     <ngx-mat-itl-tel-input placeholder="Enter phone..." [matChipInputFor]="telList" [matChipInputAddOnBlur]="true"
         (matChipInputTokenEnd)="addTel($event)" [autocomplete]="autocomplete"
