+++ projects/ngx-mat-intl-tel-input/src/lib/ngx-mat-itl-tel-input/ngx-mat-itl-tel-input.component.html
@@ -1,36 +1,8 @@
-<div class="ngx-mat-tel-input-container">
-  <button [disabled]="countrySelectionDisabled" type="button" mat-button [matMenuTriggerFor]="menu" class="country-selector">
-    <span class="country-selector-flag flag" [ngClass]="selectedCountry.flagClass"></span>
-    <span class="country-selector-code">+{{selectedCountry.dialCode}}</span>
-  </button>
-  <mat-menu #menu="matMenu">
-    <input *ngIf="enableSearch" class="country-search" [(ngModel)]="searchCriteria" type="text" placeholder="Search..."
-      (click)="$event.stopPropagation()" />
-    <button type="button" mat-menu-item class="country-list-button" *ngFor="let country of preferredCountriesInDropDown"
-      (click)="onCountrySelect(country, focusable)">
-      <div class="icon-wrapper">
-        <div class="flag" [ngClass]="country.flagClass"></div>
-      </div>
-      <div class="label-wrapper">{{country.name}} +{{country.dialCode}}</div>
-    </button>
-    <mat-divider *ngIf="preferredCountriesInDropDown?.length"></mat-divider>
-    <ng-container *ngFor="let country of allCountries">
-      <button type="button" mat-menu-item class="country-list-button" *ngIf="country.name | search:searchCriteria"
-        (click)="onCountrySelect(country, focusable)">
-        <div class="icon-wrapper">
-          <div class="flag" [ngClass]="country.flagClass"></div>
-        </div>
-        <div class="label-wrapper">{{country.name}} +{{country.dialCode}}</div>
-      </button>
-    </ng-container>
-  </mat-menu>
 
   <input *ngIf="!matChipInputFor" matInput type="tel" [autocomplete]="autocomplete" [ngClass]="cssClass" (blur)="onTouched()" [(ngModel)]="input"
-    (ngModelChange)="onPhoneNumberChange($event)" [errorStateMatcher]="errorStateMatcher" [disabled]="disabled"
-    #focusable>
+    (ngModelChange)="onPhoneNumberChange($event)" [errorStateMatcher]="errorStateMatcher" [disabled]="disabled">
 
   <input *ngIf="matChipInputFor" [matChipInputFor]="matChipInputFor" [matChipInputAddOnBlur]="matChipInputAddOnBlur"
     (matChipInputTokenEnd)="matChipInputTokenEnd.emit($event)" type="tel" [autocomplete]="autocomplete"
-    [ngClass]="cssClass" (blur)="onTouched()" [(ngModel)]="input" (ngModelChange)="onPhoneNumberChange($event)"
-    [disabled]="disabled" #focusable>
-</div>
+    [ngClass]="cssClass" (blur)="onTouched()" [(ngModel)]="input"  (ngModelChange)="onPhoneNumberChange($event)"
+    [disabled]="disabled">
