+++ projects/ngx-mat-intl-tel-input/src/lib/ngx-mat-itl-tel-input/ngx-mat-itl-tel-input.component.ts
@@ -10,14 +10,13 @@ import {
   OnInit,
   Optional,
   Output,
-  Self
+  Self,
+  ViewChild
 } from '@angular/core';
 
 import { NG_VALIDATORS, NgControl } from '@angular/forms';
-import { CountryCode, Examples } from '../data/country-code';
-import { phoneNumberValidator, validatePhoneNumber } from '../ngx-mat-itl-tel-input.validator';
-import { Country } from '../model/country.model';
-import { getExampleNumber, parsePhoneNumberFromString, PhoneNumber, AsYouType } from 'libphonenumber-js';
+import { phoneNumberValidator} from '../ngx-mat-itl-tel-input.validator';
+import { parsePhoneNumberFromString, PhoneNumber, AsYouType } from 'libphonenumber-js';
 
 import { coerceBooleanProperty } from '@angular/cdk/coercion';
 import { Subject } from 'rxjs';
@@ -29,7 +28,7 @@ import { ErrorStateMatcher } from '@angular/material/core';
   templateUrl: './ngx-mat-itl-tel-input.component.html',
   styleUrls: ['./ngx-mat-itl-tel-input.component.css'],
   providers: [
-    CountryCode,
+   
     { provide: MatFormFieldControl, useExisting: NgxMatItlTelInputComponent },
     {
       provide: NG_VALIDATORS,
@@ -78,7 +77,8 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
   }
 
   set disabled(value: boolean) {
-    this._disabled = coerceBooleanProperty(value);
+    value = coerceBooleanProperty(value);
+    this._disabled = value;
     this.stateChanges.next();
   }
 
@@ -138,7 +138,7 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
       return;
     }
 
-    this.numberInstance = typeof (value) === 'string' ? parsePhoneNumberFromString(value) : parsePhoneNumberFromString(value[value.length - 1]);
+    this.numberInstance = this.parseNumber(value);
     if (this.numberInstance) {
       const countryCode = this.numberInstance.country;
       this.input = this.numberInstance.formatNational();
@@ -146,50 +146,66 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
         return;
       }
       setTimeout(() => {
-        this.selectedCountry = this.allCountries.find(c => c.iso2 === countryCode.toLowerCase());
-        this.countryChanged.emit(this.selectedCountry);
+        this.countryCode = countryCode;
+        this.countryChanged.emit(this.countryCode);
       }, 1);
     }
 
   }
 
+  parseNumber(value) {
+    if (typeof (value) === 'string') {
+      return parsePhoneNumberFromString(value);
+    }
+    else if (Array.isArray(value)) {
+      if (value.length)
+        return parsePhoneNumberFromString(value[value.length - 1]);
+    }
+    else return undefined;
+
+
+  }
 
   ////////////////////////////////////////////////////////////////////////////////////////////////
 
 
+  @ViewChild('focusable') inputEl;
   @Input() matChipInputAddOnBlur
   @Input() matChipInputFor;
   @Output() matChipInputTokenEnd = new EventEmitter();
   @Input() autocomplete: string = "tel"
-  @Input() countrySelectionDisabled: boolean = false;
-  @Input() preferredCountries: Array<string> = [];
-  @Input() enablePlaceholder = true;
+
+
+
   @Input() cssClass;
   @Input() name: string;
-  @Input() onlyCountries: Array<string> = [];
   @Input() errorStateMatcher: ErrorStateMatcher;
-  @Input() enableSearch = this.onlyCountries.length < 10 ? false : true;
+
 
   input;
   asYouType: AsYouType = new AsYouType();
-  allCountries: Array<Country> = [];
-  preferredCountriesInDropDown: Array<Country> = [];
-  selectedCountry: Country;
-  numberInstance: PhoneNumber;
-  searchCriteria: string;
-  @Output()
-  countryChanged: EventEmitter<Country> = new EventEmitter<Country>();
 
-  static getPhoneNumberPlaceHolder(countryISOCode: any): string {
-    try {
-      return getExampleNumber(countryISOCode, Examples).number.toString();
-    } catch (e) {
-      return e;
-    }
+  
+
+  @Input()
+  get countryCode() {
+    return this._countryCode;
+  }
+  set countryCode(value: string) {
+    this._countryCode = value;
+    //@ts-ignore
+    this.asYouType = new AsYouType(value);
+    this.inputEl.focus();
   }
 
+  _countryCode: string;
+  numberInstance: PhoneNumber;
+  @Output()
+  countryChanged: EventEmitter<string> = new EventEmitter<string>();
+
+
   constructor(
-    private countryCodeData: CountryCode,
+    
     private fm: FocusMonitor,
     private elRef: ElementRef<HTMLElement>,
     @Optional() @Self() public ngControl: NgControl
@@ -201,7 +217,7 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
       this.focused = !!origin;
       this.stateChanges.next();
     });
-    this.fetchCountryData();
+
 
     if (this.ngControl != null) {
       this.ngControl.valueAccessor = this;
@@ -209,8 +225,7 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
   }
 
   ngOnInit() {
-    this.initCountryDropdown();
-    this.initSelectedCountry();
+
   }
 
 
@@ -222,7 +237,7 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
   }
 
   public onPhoneNumberChange(event): void {
-    
+
     this.asYouType.reset();
     let value;
     this.input = this.asYouType.input(event);
@@ -242,70 +257,7 @@ export class NgxMatItlTelInputComponent implements OnInit, OnDestroy, DoCheck, M
 
     console.log(value);
     this.propagateChange(value);
-  }
 
-  public onCountrySelect(country: Country, el): void {
-    this.setSelectedCountry(country);
-    el.focus();
-  }
-
-
-  protected fetchCountryData(): void {
-    this.countryCodeData.allCountries.forEach(c => {
-      const country: Country = {
-        name: c[0].toString(),
-        iso2: c[1].toString(),
-        dialCode: c[2].toString(),
-        priority: +c[3] || 0,
-        areaCodes: c[4] as string[] || undefined,
-        flagClass: c[1].toString().toUpperCase(),
-        placeHolder: ''
-      };
-
-      if (this.enablePlaceholder) {
-        country.placeHolder = NgxMatItlTelInputComponent.getPhoneNumberPlaceHolder(country.iso2.toUpperCase());
-      }
-
-      this.allCountries.push(country);
-    });
-  }
-
-  private initCountryDropdown() {
-    if (this.preferredCountries.length) {
-      this.preferredCountries.forEach(iso2 => {
-        const preferredCountry = this.allCountries.filter((c) => {
-          return c.iso2 === iso2;
-        });
-        this.preferredCountriesInDropDown.push(preferredCountry[0]);
-      });
-    }
-    if (this.onlyCountries.length) {
-      this.allCountries = this.allCountries.filter(c => this.onlyCountries.includes(c.iso2));
-    }
-
-  }
-
-  private initSelectedCountry() {
-    let country;
-    if (this.numberInstance && this.numberInstance.country) {
-      // If an existing number is present, we use it to determine selectedCountry
-      country = this.allCountries.find(c => c.iso2 === this.numberInstance.country.toLowerCase());
-    } else {
-      if (this.preferredCountriesInDropDown.length) {
-        country = this.preferredCountriesInDropDown[0];
-      } else {
-        country = this.allCountries[0];
-      }
-    }
-    this.setSelectedCountry(country);
-
-  }
-
-  setSelectedCountry(country: Country) {
-    this.selectedCountry = country;
-    //@ts-ignore
-    this.asYouType = new AsYouType(country.iso2.toUpperCase());
-    this.countryChanged.emit(this.selectedCountry);
   }
 
   reset() {
